Backend/Backend-auth/requirements.txt
# Core FastAPI and Server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Pydantic for data validation
pydantic==2.5.0
pydantic-settings==2.1.0

# OpenAI API
openai==0.28.1

# Environment variables
python-dotenv==1.0.0

# Fuzzy string matching for symptom matching
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0

# Authentication & Security
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.1
cryptography==41.0.7

# Database (if using SQLite with SQLAlchemy)
# Uncomment if needed:
# sqlalchemy==2.0.23

# Additional utilities
requests==2.31.0

# CORS middleware (already included in FastAPI, but explicit)
# No separate package needed

# For better datetime handling
python-dateutil==2.8.2